# public/gpt-actions-openapi.yaml
openapi: 3.1.0
info:
  title: kinbot API
  description: |
    kinbot（B2B営業ミーティング・コパイロット）のデータを読み取るためのREST API。
    案件・営業プロセスの抽出イベント・商談（文字起こし/要約/分析）・会社プロフィールを取得できる。
    すべて読み取り専用（GET）。認証は Bearer トークン。
  version: 1.0.0
servers:
  - url: https://kinbot-production-225f.up.railway.app
paths:
  /api/gpt/deals:
    get:
      operationId: listDeals
      summary: 案件一覧を取得
      description: >
        kinbotの案件一覧を取得する。会社名・担当者・チーム・ステータス
        （進行中 / 失注(未定) / 失注(その他) / 受注 等）・初回商談日を返す。
        管理者トークンで接続している場合、owner を省略すると全担当者分が返る。
      x-openai-isConsequential: false
      parameters:
        - name: owner
          in: query
          required: false
          description: 担当者のメールアドレスで絞り込み（任意）
          schema:
            type: string
        - name: team
          in: query
          required: false
          description: チーム名で絞り込み（任意）
          schema:
            type: string
        - name: status
          in: query
          required: false
          description: "ステータスで絞り込み（任意）例: 進行中, 失注(未定), 受注"
          schema:
            type: string
        - name: from
          in: query
          required: false
          description: 初回商談日の開始日 YYYY-MM-DD（任意）
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: false
          description: 初回商談日の終了日 YYYY-MM-DD（任意）
          schema:
            type: string
            format: date
      responses:
        "200":
          description: 案件の配列。各要素は deal_id, company_name, owner, team, status, first_meeting_date, updated_at, および company_profile（業界・従業員数employees・事業内容など。取得済みの会社のみ）を含む。
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
security:
  - bearerAuth: []